@@ -1,45 +1,46 @@
 {
   "name": "ngo-app",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
     "aos": "^3.0.0-beta.6",
-    "axios": "^0.19.0",
-    "babel-jest": "23.6.0",
-    "body-parser": "^1.19.0",
-    "bootstrap": "4.0.0",
+    "axios": "^1.6.7",
+    "babel-jest": "24.9.0",
+    "body-parser": "^1.20.2",
+    "bootstrap": "5.3.2",
     "bootstrap-social": "^5.1.1",
-    "cookie-parser": "^1.4.4",
-    "cross-fetch": "^3.0.4",
-    "debug": "^4.1.1",
-    "express": "^4.17.1",
-    "google-map-react": "^1.1.4",
+    "cookie-parser": "^1.4.6",
+    "cross-fetch": "^4.0.0",
+    "debug": "^4.3.4",
+    "express": "^4.18.2",
+    "google-map-react": "^2.2.1",
     "mdbootstrap": "^4.20.0",
-    "morgan": "^1.9.1",
-    "netlify-cli": "^12.2.0",
-    "nodemailer": "^6.3.0",
+    "morgan": "^1.10.0",
+    "netlify-cli": "^17.16.2",
+    "nodemailer": "^6.9.9",
     "popper": "^1.0.1",
-    "prop-types": "^15.7.2",
-    "pug": "^2.0.4",
-    "react": "^16.8.6",
-    "react-animated-slider": "^1.1.4",
+    "prop-types": "^15.8.1",
+    "pug": "^3.0.2",
+    "react": "^18.2.0",
+    "react-animated-components": "^4.0.0",
+    "react-animated-slider": "^2.0.0",
     "react-animation-components": "^3.0.0",
-    "react-dom": "^16.8.6",
-    "react-popper": "^1.3.3",
-    "react-redux": "^7.1.0",
+    "react-dom": "^18.0.0",
+    "react-popper": "^2.3.0",
+    "react-redux": "^9.1.0",
     "react-redux-form": "^1.16.14",
-    "react-router-dom": "^5.0.1",
-    "react-scripts": "2.1.8",
-    "react-transition-group": "^4.2.0",
-    "reactstrap": "^8.0.0",
-    "redux": "^4.0.4",
+    "react-router-dom": "^6.22.0",
+    "react-scripts": "^3.0.1",
+    "react-transition-group": "^4.4.5",
+    "reactstrap": "^9.2.2",
+    "redux": "^5.0.1",
     "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.3.0",
+    "redux-thunk": "^3.1.0",
     "serve-favicon": "^2.5.0",
-    "video-react": "^0.14.0"
+    "video-react": "^0.16.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-scripts --openssl-legacy-provider start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
@@ -62,8 +63,7 @@
   },
   "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
   "main": "index.js",
-  "devDependencies": {},
-  "homepage": "http://www.sedirwanda.org/",
+  "homepage": "https://www.sedirwanda.org/",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/TechEnfold/SediRwanda.git"
